+++ extensions/api/data-management/asset/src/main/java/org/eclipse/dataspaceconnector/api/datamanagement/asset/AssetApiExtension.java
@@ -29,10 +29,7 @@ import org.eclipse.dataspaceconnector.spi.contract.negotiation.store.ContractNeg
 import org.eclipse.dataspaceconnector.spi.system.Inject;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtension;
 import org.eclipse.dataspaceconnector.spi.system.ServiceExtensionContext;
-import org.eclipse.dataspaceconnector.spi.transaction.NoopTransactionContext;
 import org.eclipse.dataspaceconnector.spi.transaction.TransactionContext;
-
-import static java.util.Optional.ofNullable;
 
 public class AssetApiExtension implements ServiceExtension {
 
@@ -54,7 +51,7 @@ public class AssetApiExtension implements ServiceExtension {
     @Inject
     DtoTransformerRegistry transformerRegistry;
 
-    @Inject(required = false)
+    @Inject
     TransactionContext transactionContext;
 
     @Override
@@ -65,12 +62,8 @@ public class AssetApiExtension implements ServiceExtension {
     @Override
     public void initialize(ServiceExtensionContext serviceExtensionContext) {
         var monitor = serviceExtensionContext.getMonitor();
-        var transactionContextImpl = ofNullable(transactionContext)
-                .orElseGet(() -> {
-                    monitor.warning("No TransactionContext registered, a no-op implementation will be used, not suitable for production environments");
-                    return new NoopTransactionContext();
-                });
-        var service = new AssetServiceImpl(assetIndex, assetLoader, contractNegotiationStore, transactionContextImpl);
+
+        var service = new AssetServiceImpl(assetIndex, assetLoader, contractNegotiationStore, transactionContext);
 
         transformerRegistry.register(new AssetDtoToAssetTransformer());
         transformerRegistry.register(new DataAddressDtoToDataAddressTransformer());